+++ hw05_parallel_execution/go.mod
@@ -1,4 +1,4 @@
-module github.com/fixme_my_friend/hw05_parallel_execution
+module github.com/Arrgohn/otus_hw/hw05_parallel_execution
 
 go 1.16
