+++ hw07_file_copying/go.mod
@@ -1,3 +1,3 @@
-module github.com/fixme_my_friend/hw07_file_copying
+module github.com/Arrgohn/otus_hw/hw07_file_copying
 
 go 1.16
