+++ hw08_envdir_tool/go.mod
@@ -1,3 +1,3 @@
-module github.com/fixme_my_friend/hw08_envdir_tool
+module github.com/Arrgohn/otus_hw/hw08_envdir_tool
 
 go 1.16
