+++ hw09_struct_validator/go.mod
@@ -1,3 +1,3 @@
-module github.com/fixme_my_friend/hw09_struct_validator
+module github.com/Arrgohn/otus_hw/hw09_struct_validator
 
 go 1.16
