+++ hw12_13_14_15_calendar/go.mod
@@ -1,3 +1,3 @@
-module github.com/fixme_my_friend/hw12_13_14_15_calendar
+module github.com/Arrgohn/otus_hw/hw12_13_14_15_calendar
 
 go 1.16
